+++ release_notes.md
@@ -130,3 +130,4 @@ First release into griffon artifacts portal.
 ### 0.7
 1. Add `findAllModelFetch[Graph]()`, `findModelByDslFetch[Graph]()`, and `findModelByXXXFetch[Graph]()`.
 1. Fixes `@Transaction` to restrictive: allow transformation on getter and setter.
+1. Finders by default will remove duplicate entities.  Use `allowDuplicate: true` config or global `griffon.simplejpa.finders.alwaysAllowDuplicate` config to allow duplicate entities in the result.